chore: rename relation to relationtuple consistently
